backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
langchain-huggingface
langchain-text-splitters
sentence-transformers
faiss-cpu
rank-bm25
pymupdf
python-docx
ebooklib
beautifulsoup4
python-dotenv
python-multipart
numpy
# Auth
python-jose[cryptography]
passlib[bcrypt]
# Database
sqlalchemy
psycopg2-binary
